backend/requirements.txt
Django==2.2.19
djoser==2.1.0
django-colorfield==0.8.0
django-cors-headers
gunicorn==20.1.0
psycopg2-binary==2.9.5
python-dotenv==0.21.1
Pillow==9.3.0
requests==2.26.0
djangorestframework==3.12.4
drf-extra-fields==3.4.1
PyJWT==2.1.0
